requirements.txt
ultralytics==8.0.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.18.0
pandas>=1.1.0
scipy>=1.5.0
onnx>=1.8.0
onnxruntime>=1.7.0
streamlit>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.50.0
tensorboard>=2.10.0
pyyaml>=5.4.0
requests>=2.25.0
